requirements.txt
flask
openai
ffmpeg-python
openai-whisper
pandas
numpy
scikit-learn
joblib
streamlit
torch
gunicorn
transformers
sentence-transformers
faiss-cpu